Refatorado Services(user, menu)

+++ src/services/user.service.js
@@ -1,9 +1,10 @@
 import UserRepository from '../repositories/user.repository.js'
 import { UserResponseDTO } from '../dtos/user.dto.js'
+import bcrypt from "bcrypt"
 
 //Serviço da rota de usuários, com criação, alteração, exposição e como deletar os usuários.
 class UserService {
-    static async create(createUserData){
+    async create(createUserData){
 
         const email = await UserRepository.findByEmail(createUserData.email)
         
@@ -13,18 +14,26 @@ class UserService {
             throw error
         }
         
+        //Hashing e alteração de senha do userData
+        const salt = await bcrypt.genSalt(10)
+        const hashedSenha = await bcrypt.hash(createUserData.senha, salt)
+
+        createUserData.senha = hashedSenha
+
         const newUserFromDb = await UserRepository.create(createUserData)
+        const newUserDto = UserResponseDTO(newUserFromDb)
 
-        return new UserResponseDTO(newUserFromDb)
+        return newUserDto
 
     }
 
-    static async getAll(){
-        const users = await UserRepository.getAll()
-        return users.map((user)=>new UserResponseDTO(user))
+    async getAll(){
+        const userFromRepo = await UserRepository.findAll()
+        const userDTO = UserResponseDTO(userFromRepo)
+        return userDTO
     }
 
-    static async update(id, updateUserData){
+    async update(id, updateUserData){
         
         const user = await UserRepository.findById(id)
 
@@ -33,12 +42,31 @@ class UserService {
             error.statusCode = 404
             throw error
         }
+
+        //Impedindo duplicação de email, caso tenha atualizado
+        if (updateUserData.email && updateUserData.email !== user.email) {
+            const existing = await UserRepository.findByEmail(updateUserData.email)
+
+            if (existing) {
+                const error = new Error('Este email já está em uso.')
+                error.statusCode = 409
+                throw error;
+            }
+        }
+
+        //Hashing da senha, caso tenha atualizado
+        if(updateUserData.senha){
+            const salt = await bcrypt.genSalt(10)
+            const hashedSenha = await bcrypt.hash(updateUserData.senha, salt)
+            updateUserData.senha = hashedSenha
+        }
         
         const updateUserFromDb = await UserRepository.update(id, updateUserData)
-        return new UserResponseDTO(updateUserFromDb)
+        const updateUserDto = UserResponseDTO(updateUserFromDb)
+        return updateUserDto
     }
 
-    static async getById(id){
+    async getById(id){
         const user = await UserRepository.findById(id)
 
         if(!user){
@@ -47,21 +75,23 @@ class UserService {
             throw error
         }
 
-        return new UserResponseDTO(user)
+        const userDTO = UserResponseDTO(user)
+
+        return userDTO
     }
 
-    static async delete(id){
+    async delete(id){
         
-        const deletedUser = await UserRepository.delete(id)
-
-        if(!deletedUser){
-            const error = new Error('Usuário não encontrado')
-            error.statusCode = 404
-            throw error
+        const user = await UserRepository.findById(id);
+        if (!user) {
+            const error = new Error('Usuário não encontrado.');
+            error.statusCode = 404;
+            throw error;
         }
-
-        return new UserResponseDTO(deletedUser)
+        
+        return await UserRepository.delete(id);
     }
+    
 }
 
-export default UserService
+export default new UserService()
